added u.id in filter SQL query
